Refactor the organization of GitHub components
